refactor(debt): relocate server lifecycle to runtime/server.rs (D-01 step 2a)
Verbatim move of start_embedded_server + check_local_health + read_server_logs
+ stop_local_server + SERVER_PROCESS/SERVER_START_LOCK/server_start_lock into a
new runtime/server.rs submodule (use super::*; commands re-exported for the
android-only generate_handler!). Shared helpers (runtime_dir, native_lib_dir,
force_symlink, prepare_toolchain_wrappers, check_health) stay in runtime.rs.

runtime.rs 1492 -> 745 LOC. Done via a deterministic anchor-based split script
(no hand transcription). Verified WSL: cargo test --lib runtime 14/0, clippy
--lib --tests -D warnings clean. Decomposition of the function follows (2b).
